added roles tab to clients
